+++ src/frontend/app/shared-tdoc/components/tdoc-list-item/tdoc-list-item.component.ts
@@ -1,7 +1,10 @@
-import {ChangeDetectionStrategy, ChangeDetectorRef, Component} from '@angular/core';
+import {ChangeDetectionStrategy, ChangeDetectorRef, Component, EventEmitter, Input, Output} from '@angular/core';
 import {LayoutService} from '@dps/mycms-frontend-commons/dist/angular-commons/services/layout.service';
 import {CommonDocListItemComponent} from '@dps/mycms-frontend-commons/dist/frontend-cdoc-commons/components/cdoc-list-item/cdoc-list-item.component';
 import {TourDocContentUtils} from '../../services/tdoc-contentutils.service';
+import {CommonDocRecord} from '@dps/mycms-commons/dist/search-commons/model/records/cdoc-entity-record';
+import {TourDocRecord} from '../../../../shared/tdoc-commons/model/records/tdoc-record';
+import {PageUtils} from '@dps/mycms-frontend-commons/dist/angular-commons/services/page.utils';
 
 @Component({
     selector: 'app-tdoc-list-item',
@@ -10,8 +13,34 @@ import {TourDocContentUtils} from '../../services/tdoc-contentutils.service';
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class TourDocListItemComponent  extends CommonDocListItemComponent {
+    protected mapFlagSymbol = '&#128681';
+    protected mapFlagAvailable = false;
+
+    @Input()
+    public showItemMapFlag?: false;
+
+    @Output()
+    public showItemOnMap: EventEmitter<CommonDocRecord> = new EventEmitter();
+
     constructor(contentUtils: TourDocContentUtils, cd: ChangeDetectorRef, layoutService: LayoutService) {
-        super(contentUtils, cd, layoutService);
+        super(contentUtils, cd, layoutService, );
         this.listLayoutName = 'default';
     }
+
+    protected updateData() {
+        const tdocRecord: TourDocRecord = <TourDocRecord>this.record;
+        if (tdocRecord && this.showItemMapFlag && (tdocRecord.geoLat || tdocRecord.gpsTrackSrc || tdocRecord.gpsTrackBasefile)) {
+            this.mapFlagAvailable = true;
+        } else {
+            this.mapFlagAvailable = false;
+        }
+        super.updateData();
+        this.cd.markForCheck();
+    }
+
+    public submitShowItemOnMap(tdoc: TourDocRecord) {
+        this.showItemOnMap.emit(tdoc);
+        return false;
+    }
+
 }