+++ src/shared/tdoc-commons/services/tdoc-sql-mytb.adapter.ts
@@ -62,5 +62,22 @@ export class TourDocSqlMytbAdapter extends GenericSqlAdapter<TourDocRecord, Tour
 
         return undefined;
     }
+
+    protected transformToSqlDialect(sql: string): string {
+        if (this.config.knexOpts.client !== 'mysql') {
+            // dirty workaround because sqlite has no functions as mysql
+            sql = sql.replace(/GetLocationNameAncestry\(location.l_id, location.l_name, " -> "\)/g,
+                '"T" || location.l_typ || "L" || location.l_parent_id || " -> " || location.l_name');
+            sql = sql.replace(/GetLocationIdAncestry\(location.l_id, ","\)/g,
+                'CAST(location.l_parent_id AS CHAR(50)) || "," || CAST(location.l_id AS CHAR(50))');
+            sql = sql.replace('CONCAT(CAST(location.l_parent_id AS CHAR(50)), ",", CAST(location.l_id AS CHAR(50)))',
+                'CAST(location.l_parent_id AS CHAR(50)) || "," || CAST(location.l_id AS CHAR(50))');
+            sql = sql.replace(/GetTechName\(([a-zA-Z0-9_.]+)\)/g,
+                'REPLACE(REPLACE(LOWER($1), " ", "_"), "/", "_")');
+        }
+
+        return super.transformToSqlDialect(sql);
+    }
+
 }